feat: Enable selective import of AppMap capabilities
